requirements.txt
# Core FastAPI and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6

# OpenAI and LLM
openai>=1.0.0
tiktoken>=0.5.0

# LangChain ecosystem
langchain>=0.0.300
langchain-openai>=0.0.1
langchain-community>=0.0.1

# Vector store and embeddings
faiss-cpu>=1.7.0
numpy>=1.20.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# HTTP requests and file handling
httpx>=0.24.0
aiofiles>=23.0.0
requests>=2.28.0

# Text processing and utilities
python-dotenv>=1.0.0
typing-extensions>=4.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Additional utilities (built-in modules don't need to be listed)
# pathlib, tempfile, os, and logging are built-in Python modules
